requirements/base.txt
Django==1.7.1
#-e git+git://github.com/django/django.git@1.7c2#egg=django
#South==0.8.1
argparse==1.2.1
django-debug-toolbar==1.0.1
django-suit==0.2.11
geojson==1.0.1
#newrelic==2.10.0.8
psycopg2==2.5.1
raven==3.5.0
sqlparse==0.1.10
wsgiref==0.1.2
django_extensions==1.3.3
Pillow
newrelic==2.32.0.28

#pygments, markdown and django-filter are optional dependencies for djangorestfamework
djangorestframework==2.3.14
pygments==1.6
markdown==2.4.1
django-filter==0.7

#https://pypi.python.org/pypi/drf-compound-fields/0.2.0
drf-compound-fields==0.2.1

#https://github.com/mjumbewu/django-rest-framework-csv
djangorestframework-csv==1.3.3

#https://github.com/djangonauts/django-rest-framework-gis
djangorestframework-gis==0.2

#https://github.com/marcgibbons/django-rest-swagger
django-rest-swagger

#http://django-compressor.readthedocs.org/ (for js / css minification)
django-compressor==1.4

#https://github.com/Hipo/drf-extra-fields (for Base64ImageField)
django-extra-fields==0.1

python-dateutil==2.3
pytz==2014.10
requests==2.5
